+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file is similar to the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## Unreleased
+- [Patch] Remove onFocus update to input value in Autocomplete
 
 ## 40.11.1 - 2018-11-01
 - [Patch] Include src/ in package to enable external bundling